# AV2/bazar-tem-tudo/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs do BazarTemTudo",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/enderecos": {
            "get": {
                "tags": [
                    "Enderecos"
                ],
                "summary": "Lista todos os endereços",
                "operationId": "3187938c343b2d7b76ed175ecba9b417",
                "responses": {
                    "200": {
                        "description": "Lista de endereços",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "nivel_servico_envio": {
                                                "type": "string",
                                                "example": "Express"
                                            },
                                            "endereco_envio_linha1": {
                                                "type": "string",
                                                "example": "Rua A, 123"
                                            },
                                            "endereco_envio_linha2": {
                                                "type": "string",
                                                "example": "Apt 101"
                                            },
                                            "endereco_envio_linha3": {
                                                "type": "string",
                                                "example": "Bloco B"
                                            },
                                            "cidade_envio": {
                                                "type": "string",
                                                "example": "São Paulo"
                                            },
                                            "estado_envio": {
                                                "type": "string",
                                                "example": "SP"
                                            },
                                            "codigo_postal_envio": {
                                                "type": "string",
                                                "example": "12345-678"
                                            },
                                            "pais_envio": {
                                                "type": "string",
                                                "example": "Brasil"
                                            },
                                            "id_cliente": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string"
                                            },
                                            "updated_at": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Enderecos"
                ],
                "summary": "Criar um novo endereço",
                "operationId": "4554fd24c6358cd5168606ab307b0360",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nivel_servico_envio": {
                                        "type": "string",
                                        "example": "Express"
                                    },
                                    "endereco_envio_linha1": {
                                        "type": "string",
                                        "example": "Rua A, 123"
                                    },
                                    "endereco_envio_linha2": {
                                        "type": "string",
                                        "example": "Apt 101"
                                    },
                                    "endereco_envio_linha3": {
                                        "type": "string",
                                        "example": "Bloco B"
                                    },
                                    "cidade_envio": {
                                        "type": "string",
                                        "example": "São Paulo"
                                    },
                                    "estado_envio": {
                                        "type": "string",
                                        "example": "SP"
                                    },
                                    "codigo_postal_envio": {
                                        "type": "string",
                                        "example": "12345-678"
                                    },
                                    "pais_envio": {
                                        "type": "string",
                                        "example": "Brasil"
                                    },
                                    "id_cliente": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Endereço criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "nivel_servico_envio": {
                                            "type": "string",
                                            "example": "Express"
                                        },
                                        "endereco_envio_linha1": {
                                            "type": "string",
                                            "example": "Rua A, 123"
                                        },
                                        "endereco_envio_linha2": {
                                            "type": "string",
                                            "example": "Apt 101"
                                        },
                                        "endereco_envio_linha3": {
                                            "type": "string",
                                            "example": "Bloco B"
                                        },
                                        "cidade_envio": {
                                            "type": "string",
                                            "example": "São Paulo"
                                        },
                                        "estado_envio": {
                                            "type": "string",
                                            "example": "SP"
                                        },
                                        "codigo_postal_envio": {
                                            "type": "string",
                                            "example": "12345-678"
                                        },
                                        "pais_envio": {
                                            "type": "string",
                                            "example": "Brasil"
                                        },
                                        "id_cliente": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}